added authorization

+++ routes.php
@@ -4,19 +4,21 @@
 
 $router->get('/', 'HomeController@index');
 $router->get('/listings', 'ListingController@index');
-$router->get('/listings/create', 'ListingController@create');
-$router->get('/listings/edit/{id}', 'ListingController@edit');
+$router->get('/listings/create', 'ListingController@create', ['auth']);
+$router->get('/listings/edit/{id}', 'ListingController@edit', ['auth']);
 $router->get('/listings/{id}', 'ListingController@show');
 
 // POST request
-$router->post('/listings', 'ListingController@store');
-$router->put('/listings/{id}', 'ListingController@update');
-$router->delete('/listings/{id}', 'ListingController@destroy');
+$router->post('/listings', 'ListingController@store', ['auth']);
+$router->put('/listings/{id}', 'ListingController@update', ['auth']);
+$router->delete('/listings/{id}', 'ListingController@destroy', ['auth']);
 
 //Users controller
 
-$router->get('/auth/create', 'UserController@create');
-$router->get('/auth/login', 'UserController@login');
+$router->get('/auth/create', 'UserController@create', ['guest']);
+$router->get('/auth/login', 'UserController@login', ['guest']);
 
-$router->post('/auth/register', 'UserController@store');
+$router->post('/auth/register', 'UserController@store', ['guest']);
+$router->post('/auth/logout', 'UserController@logout', ['auth']);
+$router->post('/auth/login', 'UserController@authenticate', ['guest']);
